+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/CompoundAction.java
@@ -115,6 +115,11 @@ public abstract class CompoundAction extends BaseSpellAction
             }
 
             result = result.min(step(context));
+
+            // Prevent infinite loops of no actions
+            if (context.getWorkAllowed() > 0 && !result.isStop()) {
+                result = SpellResult.PENDING;
+            }
         }
         return result;
     }